+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+- Settings menus.
+    - Enable Validation Automator
+    - Enable Texture Format Check on Standalone
+
+### Changed
+- [NDMF] No longer check texture formats on Windows build target by default.
+
 ## [2.10.1] - 2025-06-02
 
 ### Fixed